1.optimize the performance of the birt reports 2.revert BirtReportsUploadAction.java
git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@11907 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ mifos/src/org/mifos/application/reports/business/service/ReportsDataService.java
@@ -1,5 +1,6 @@
 package org.mifos.application.reports.business.service;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -82,7 +83,7 @@ public class ReportsDataService {
 						convertIntegerToShort(loanProductId));
 	}
 
-	public int getTotalOutstandingPrincipalOfLoanAccountsInActiveGoodStanding(
+	public BigDecimal getTotalOutstandingPrincipalOfLoanAccountsInActiveGoodStanding(
 			Integer branchId, Integer loanOfficerId, Integer loanProductId)
 			throws PersistenceException {
 		return loanPersistence
